+++ crates/vgonio-comp/src/app/cache.rs
@@ -488,25 +488,40 @@ impl RawCache {
             log::debug!("  Refractive index database not found at {:?}", sys_path);
         }
 
+        let excluded = match &config.user.excluded_ior_files {
+            Some(excluded) => excluded.as_slice(),
+            None => &[],
+        };
+
         // First load refractive indices from `VgonioConfig::sys_data_dir()`.
-        let n = Self::load_refractive_indices(&mut self.iors, &sys_path);
+        let n = Self::load_refractive_indices(&mut self.iors, &sys_path, excluded);
         log::debug!("  Loaded {} ior files from {:?}", n, sys_path);
 
         if !user_path.as_ref().is_some_and(|p| p.exists()) {
             log::debug!("  Refractive index database not found at {:?}", user_path);
         }
 
-        let n = Self::load_refractive_indices(&mut self.iors, user_path.as_ref().unwrap());
+        let n =
+            Self::load_refractive_indices(&mut self.iors, user_path.as_ref().unwrap(), excluded);
         log::debug!("  Loaded {} ior files from {:?}", n, user_path);
     }
 
     // TODO: move to RefractiveIndexRegistry
     /// Load the refractive index database from the given path.
     /// Returns the number of files loaded.
-    fn load_refractive_indices(iors: &mut RefractiveIndexRegistry, path: &Path) -> u32 {
+    fn load_refractive_indices(
+        iors: &mut RefractiveIndexRegistry,
+        path: &Path,
+        excluded: &[String],
+    ) -> u32 {
         let mut n_files = 0;
         if path.is_file() {
             log::debug!("Loading refractive index database from {:?}", path);
+            let filename = path.file_name().unwrap().to_str().unwrap();
+            if excluded.contains(&filename.to_string()) {
+                log::debug!("  -- excluded: {}", filename);
+                return 0;
+            }
             let medium = Medium::from_str(
                 path.file_name()
                     .unwrap()
@@ -531,7 +546,7 @@ impl RawCache {
             for entry in path.read_dir().unwrap() {
                 let entry = entry.unwrap();
                 let path = entry.path();
-                n_files += Self::load_refractive_indices(iors, &path);
+                n_files += Self::load_refractive_indices(iors, &path, excluded);
             }
         }